+++ CODEFORCES/1100/1854A1.cpp
@@ -1,48 +1,42 @@
-#include <iostream>
-#include <vector>
-
+#include <bits/stdc++.h>
 using namespace std;
-
-vector<pair<int, int>> make_non_decreasing(int n, vector<int>& arr) {
-    vector<pair<int, int>> operations;
-    for (int i = 0; i < n - 1; ++i) {
-        while (arr[i] > arr[i + 1]) {
-            for (int j = i + 1; j < n; ++j) {
-                if (arr[i] >= arr[j]) {
-                    operations.push_back(make_pair(i + 1, j + 1));
-                    arr[i] += arr[j];
-                    break;
-                }
+int main(){
+    int T; cin >> T;
+    while(T--){
+        int n; cin >> n;
+        vector<int> a(n);
+        int p = -1;
+        for(int i = 0; i < n; i++){
+            cin >> a[i];
+            if(a[i] > 0){
+                p = i;
             }
-            // If no j satisfies arr[i] >= arr[j], break the while loop
-            if (arr[i] <= arr[i + 1])
-                break;
+        }
+        if(n==1){
+            cout << 0 << endl;
+            continue;
+        }
+        if(p==-1){
+            cout << n-1 << endl;
+            for(int i = 0; i < n-1; i++){
+                cout << n-i-1 << " " << n-i <<endl;
+            }
+        }else{
+            vector<pair<int,int>> ops;
+            while(a[p] < 20){
+                a[p] *= 2;
+                ops.push_back({p+1,p+1});
+            }
+            ops.push_back({2, p+1});
+            ops.push_back({2, p+1});
+            for(int i = 3; i <= n; i++){
+                ops.push_back({i, i-1});
+                ops.push_back({i, i-1});
+            }
+            cout << ops.size() << endl;
+            for(auto [x,y]: ops){
+                cout << x << " " << y << endl;
+            }
         }
     }
-    return operations;
-}
-
-int main() {
-    int t; // Number of test cases
-    cin >> t;
-
-    while (t--) {
-        int n; // Length of the array
-        cin >> n;
-        vector<int> arr(n);
-
-        for (int i = 0; i < n; ++i)
-            cin >> arr[i];
-
-        vector<pair<int, int>> operations = make_non_decreasing(n, arr);
-        int k = operations.size();
-
-        // Output the result
-        cout << k << "\n";
-        for (auto op : operations) {
-            cout << op.first << " " << op.second << "\n";
-        }
-    }
-
-    return 0;
 }